changed dao module from subscriptions to subscription_parts
